+++ packages/mistralai-gcp/src/funcs/fimComplete.ts
@@ -83,6 +83,7 @@ export async function fimComplete(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,