+++ packages/mistralai-gcp/src/models/components/chatcompletionstreamrequest.ts
@@ -303,7 +303,7 @@ export const ChatCompletionStreamRequest$inboundSchema: z.ZodType<
 > = z.object({
   model: z.nullable(z.string()),
   temperature: z.nullable(z.number()).optional(),
-  top_p: z.number().default(1),
+  top_p: z.number().optional(),
   max_tokens: z.nullable(z.number().int()).optional(),
   stream: z.boolean().default(true),
   stop: z.union([z.string(), z.array(z.string())]).optional(),
@@ -355,7 +355,7 @@ export const ChatCompletionStreamRequest$inboundSchema: z.ZodType<
 export type ChatCompletionStreamRequest$Outbound = {
   model: string | null;
   temperature?: number | null | undefined;
-  top_p: number;
+  top_p?: number | undefined;
   max_tokens?: number | null | undefined;
   stream: boolean;
   stop?: string | Array<string> | undefined;
@@ -382,7 +382,7 @@ export const ChatCompletionStreamRequest$outboundSchema: z.ZodType<
 > = z.object({
   model: z.nullable(z.string()),
   temperature: z.nullable(z.number()).optional(),
-  topP: z.number().default(1),
+  topP: z.number().optional(),
   maxTokens: z.nullable(z.number().int()).optional(),
   stream: z.boolean().default(true),
   stop: z.union([z.string(), z.array(z.string())]).optional(),